+++ src/lib/FlightTasks/tasks/Auto/FlightTaskAuto.cpp
