+++ src/lib/FlightTasks/tasks/Auto/FlightTaskAuto.hpp
