+++ src/lib/FlightTasks/tasks/AutoLine/FlightTaskAutoLine.cpp
