+++ src/lib/FlightTasks/tasks/ManualAltitudeSmooth/FlightTaskManualAltitudeSmooth.cpp
