+++ src/lib/FlightTasks/tasks/ManualAltitudeSmooth/FlightTaskManualAltitudeSmooth.hpp
