+++ src/lib/FlightTasks/tasks/ManualPosition/FlightTaskManualPosition.cpp
