+++ src/lib/FlightTasks/tasks/ManualPositionSmooth/FlightTaskManualPositionSmooth.hpp
