+++ src/lib/FlightTasks/tasks/ManualStabilized/FlightTaskManualStabilized.cpp
