+++ src/lib/FlightTasks/tasks/ManualStabilized/FlightTaskManualStabilized.hpp
